+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Escape characters in query strings now work as expected, e.g. `\n` will be treated as a newline character instead of the literal string `\n`.
+- Fixed a typo in the `<=` token that could cause a display issue.
 
 ## [v3.4.0] - 2026-03-19